requirements.txt
azure-functions
streamlit
requests
msal
python-dotenv
google-generativeai
azure-data-tables